WIP: First cut of conversionLoss plugins
