+++ scijava/scijava-ops/src/main/java/module-info.java
@@ -7,6 +7,7 @@ module org.scijava.ops {
 	exports org.scijava.ops.core.builder; // contains OpBuilder classes
 	exports org.scijava.ops.matcher;
 	exports org.scijava.ops.simplify;
+	exports org.scijava.ops.conversionLoss;
 	// TODO: move OpWrapper to its own package (org.scijava.ops.wrap??)
 	exports org.scijava.ops.util; // contains OpWrapper interface
 	exports org.scijava.struct;
