+++ scijava/scijava-ops/src/main/java/org/scijava/ops/conversionLoss/LossCalculation.java
@@ -1,0 +1,41 @@
+
+package org.scijava.ops.conversionLoss;
+
+/**
+ * Container storing the result of converion loss between types {@code T} and
+ * {@code R}. Conversion loss is stored for <b>both directions of
+ * conversion</b>.
+ * 
+ * @author Gabriel Selzer
+ * @param <T> - the first type
+ * @param <R> - the second type
+ */
+public class LossCalculation<T, R> {
+
+	// TODO: maybe these should be Nils?
+	final T fromType;
+	final R toType;
+	final double fLoss;
+	final double rLoss;
+
+	public LossCalculation(T from, R to, double fLoss, double rLoss) {
+		this.fromType = from;
+		this.toType = to;
+		this.fLoss = fLoss;
+		this.rLoss = rLoss;
+	}
+
+	public double forwardLoss() {
+		return fLoss;
+	}
+
+	public double reverseLoss() {
+		return rLoss;
+	}
+
+	@Override
+	public String toString() {
+		return "Forward Loss: " + fLoss + ", Reverse Loss: " + rLoss;
+	}
+
+}
